RPi/requirements.txt
azure-storage-blob
numpy
pyttsx3
"picamera[array]"
opencv-contrib-python==4.1.0.25